requirements-test.txt
-r requirements.txt
aiopurpleair==2025.08.1
pytest-homeassistant-custom-component==0.13.318
mypy>=1.11
